Made IJmesPathGenerator independant from the parser.
# Conflicts:
#	jmespath.net.sln
#	src/jmespath.net.parser/jmespath.net.parser.csproj
